requirements.txt
1.base64
2.re=2.2.1
3.cv2=4.1.2
4.flask

